Refactor: Rename metrics to internal/telemetry
