+++ internal/telemetry/sync_gauge.go
@@ -1,4 +1,4 @@
-package metrics
+package telemetry
 
 import (
 	"context"
